refactor: change folder structure + create merged component flowEditor
